consolidate duplicated code
